requirements.txt
streamlit
openai
jinja2
weasyprint
pyyaml